Prevent users from adding contents to file dictionary with "contents" argument
For `mahautils.dictionary.Dictionary` objects, it is possible to add contents to the dictionary upon initialization with the "contents" keyword argument. However, the `mahautils.dictionary.FileDictionary` class performs checks and conversions before adding items to the dictionary, so it is undesirable to allow users to directly add content at initialization. Therefore, this option was disabled