Fix function to strip matched parentheses so that only matched parentheses pairs are removed
Previously, this function would remove a leading/trailing parenthesis pair; however, it should only perform this step if the parentheses to be removed form an open/closed pair. For instance, the leading/trailing parentheses in "((text)())" should be removed, but not the leading/trailing parentheses in "()text()" (because in the second case, there are two distinct parenthesis pairs, one before "text" and one after). Therefore, additional checks were added to ensure that only complete parenthesis pairs were stripped