Create `mahautils.generic_objects` module and `TypedList` class
Created a new module to store objects useful to a range of other code. Added a new class, `mahautils.generic_objects.TypedList` which provides almost all the functionality of a Python `list`, but requires that all items be of a specific, user-defined type. Also added tests for this new class